requirements-web.txt
flask>=2.0.0
gunicorn==21.2.0
edge-tts>=7.2.3
python-dotenv>=1.0.1
